Changed variable upperPos to DWORD, adopted from MAME 0.105 to avoid GCC 4.4 compiler warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@2909 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/windows/fileio.c
@@ -453,7 +453,7 @@ osd_file *osd_fopen(int pathtype, int pathindex, const char *filename, const cha
 {
 	DWORD disposition = 0, access = 0, sharemode = 0;
 	TCHAR fullpath[1024];
-	LONG upperPos = 0;
+	DWORD upperPos = 0;	// adopted from MAME 0.105
 	osd_file *file;
 	int i;
